+++ tests/Course/Dummy/UserTest.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Solital\Test\Course\Dummy;
+
+class UserTest
+{
+    public function testLow()
+    {
+        return "Running Low...";
+    }
+
+    public function testNormal()
+    {
+        return "Running Normal...";
+    }
+
+    public function testHigh()
+    {
+        return "Running High...";
+    }
+}